backend/pyproject.toml
[project]
name = "dynamic-crud-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "alembic>=1.14.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
